Move error types to where the are returned
