+++ nym-vpn-lib/src/credentials/mod.rs
@@ -2,10 +2,13 @@ mod check;
 mod helpers;
 mod import;
 
-mod error;
-
 pub use check::{
-    check_credential_base58, check_credential_file, check_imported_credential, check_raw_credential,
+    check_credential_base58, check_credential_file, check_imported_credential,
+    check_raw_credential, CheckImportedCredentialError,
 };
-pub use error::{CheckImportedCredentialError, CredentialStoreError, ImportCredentialError};
-pub use import::{import_credential, import_credential_base58, import_credential_file};
+pub use helpers::{
+    CredentialCoconutApiClientError, CredentialNyxdClientError, CredentialStoreError,
+};
+pub use import::{
+    import_credential, import_credential_base58, import_credential_file, ImportCredentialError,
+};
